requirements.txt
streamlit==1.31.0
pandas==2.2.0
plotly==5.18.0
supabase==1.2.0
python-dotenv==1.0.0
numpy==1.26.0